+++ packages/builder-tools/connect-studio/helpers.ts
@@ -2,6 +2,18 @@ import { PowerhouseConfig } from "@powerhousedao/config/powerhouse";
 import fs from "node:fs";
 import { join, resolve } from "node:path";
 
+export function copyConnect(sourcePath: string, targetPath: string) {
+  try {
+    // Ensure targetPath is removed before copying
+    fs.rmSync(targetPath, { recursive: true, force: true });
+
+    // Copy everything from sourcePath to targetPath
+    fs.cpSync(sourcePath, targetPath, { recursive: true });
+  } catch (error) {
+    console.error(`❌ Error copying ${sourcePath} to ${targetPath}:`, error);
+  }
+}
+
 export function backupIndexHtml(appPath: string, restore = false) {
   const filePath = join(appPath, "index.html");
   const backupPath = join(appPath, "index.html.bak");
@@ -36,6 +48,7 @@ export function removeBase64EnvValues(appPath: string) {
         `"LOCAL_DOCUMENT_EDITORS": ""`,
       );
 
+    console.log("Modified data:", modifiedData);
     // Write the modified content back to the file
     fs.writeFile(filePath, modifiedData, "utf-8", (err) => {
       if (err) {