+++ packages/design-system/src/index.ts
@@ -1,4 +1,4 @@
-import "./globals.css";
+import "../style.css";
 export * from "./connect/index.js";
 export * from "./powerhouse/index.js";
 export * from "./rwa/index.js";