+++ packages/design-system/src/scalars/components/date-time-field/date-time-picker-contet.tsx
@@ -1,5 +1,5 @@
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@radix-ui/react-tabs";
-import { Icon } from "index";
+import { Icon } from "#powerhouse";
 import { cn } from "../../lib/utils";
 import { Calendar } from "../date-picker-field/subcomponents/calendar/calendar";
 import { SelectBaseProps } from "../enum-field/types";