+++ clis/ph-cli/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.35.6-dev.0 (2025-03-08)
+
+### 🧱 Updated Dependencies
+
+- Updated @powerhousedao/connect to 1.0.0-dev.219
+
+## 0.35.5 (2025-03-08)
+
+### 🩹 Fixes
+
+- **ph-cmd:** build with tsc ([26965361](https://github.com/powerhouse-inc/powerhouse/commit/26965361))
+- **ph-cli:** cleanup dependencies ([c1fb1b63](https://github.com/powerhouse-inc/powerhouse/commit/c1fb1b63))
+- **switchboard:** updated sky-ph dependency ([39ddbfdf](https://github.com/powerhouse-inc/powerhouse/commit/39ddbfdf))
+
+### ❤️ Thank You
+
+- acaldas @acaldas
+
 ## 0.35.5-dev.0 (2025-03-06)
 
 ### 🧱 Updated Dependencies